+++ packages/ub-pub/locale/lang-ky.js
@@ -51,7 +51,7 @@ UB.i18nExtend({
   keyDevise: 'Название носителя',
   'Selected certificate': 'Выбранный сертификат',
   'Select certificate file': 'Выберите файлы сертификатов',
-  'Load certificates': 'К сожалению, не удалось найти сертификат. Выберите и загрузите ваш сертификат, который соответствует личному ключу.',
+  LoadCertificatesFails: 'LoadCertificatesFails',
   'Load': 'Загрузить',
   'Cancel': 'Отменить',
   'ByKeyFile': 'Файл',